+++ uds/translator/data_record_definitions/__init__.py
@@ -1,7 +1,21 @@
 """Data Records definitions for basic Translator."""
 
 from .conditional import (
+    CONDITIONAL_CERTIFICATE_CLIENT,
+    CONDITIONAL_CERTIFICATE_DATA,
+    CONDITIONAL_CERTIFICATE_SERVER,
+    CONDITIONAL_CHALLENGE_CLIENT,
+    CONDITIONAL_CHALLENGE_SERVER,
     CONDITIONAL_EVENT_TYPE_RECORD_09_2020,
+    CONDITIONAL_MEMORY_ADDRESS_AND_SIZE,
+    CONDITIONAL_OPTIONAL_ADDITIONAL_PARAMETER,
+    CONDITIONAL_OPTIONAL_CHALLENGE_CLIENT,
+    CONDITIONAL_OPTIONAL_EPHEMERAL_PUBLIC_KEY_CLIENT,
+    CONDITIONAL_OPTIONAL_EPHEMERAL_PUBLIC_KEY_SERVER,
+    CONDITIONAL_OPTIONAL_NEEDED_ADDITIONAL_PARAMETER,
+    CONDITIONAL_OPTIONAL_SESSION_KEY_INFO,
+    CONDITIONAL_PROOF_OF_OWNERSHIP_CLIENT,
+    CONDITIONAL_PROOF_OF_OWNERSHIP_SERVER,
     DTC_SNAPSHOT_RECORDS_LIST_2013,
     DTC_SNAPSHOT_RECORDS_LIST_2020,
     DTC_STORED_DATA_RECORDS_LIST_2013,
@@ -56,14 +70,14 @@ from .dtc import (
     FAULT_DETECTION_COUNTER,
     GROUP_OF_DTC,
     MULTIPLE_DTC_AND_FAULT_DETECTION_COUNTER,
-    MULTIPLE_DTC_SEVERITY_FUNCTIONAL_UNIT_DTC_AND_STATUS,
-    MULTIPLE_DTC_SEVERITY_DTC_AND_STATUS,
     MULTIPLE_DTC_AND_SNAPSHOT_RECORD_NUMBER,
     MULTIPLE_DTC_AND_STATUS,
-    OPTIONAL_DTC_SEVERITY_FUNCTIONAL_UNIT_DTC_AND_STATUS,
+    MULTIPLE_DTC_SEVERITY_DTC_AND_STATUS,
+    MULTIPLE_DTC_SEVERITY_FUNCTIONAL_UNIT_DTC_AND_STATUS,
     OPTIONAL_DTC_AND_STATUS,
     OPTIONAL_DTC_EXTENDED_DATA_RECORD_NUMBER,
     OPTIONAL_DTC_EXTENDED_DATA_RECORDS_DATA_LIST,
+    OPTIONAL_DTC_SEVERITY_FUNCTIONAL_UNIT_DTC_AND_STATUS,
 )
 from .other import (
     ADDITIONAL_PARAMETER_LENGTH,
@@ -80,11 +94,6 @@ from .other import (
     CHALLENGE_CLIENT_LENGTH,
     CHALLENGE_SERVER_LENGTH,
     COMMUNICATION_CONFIGURATION,
-    CONDITIONAL_CERTIFICATE_CLIENT,
-    CONDITIONAL_CERTIFICATE_DATA,
-    CONDITIONAL_CERTIFICATE_SERVER,
-    CONDITIONAL_CHALLENGE_CLIENT,
-    CONDITIONAL_CHALLENGE_SERVER,
     CONDITIONAL_COMMUNICATION_CONTROL_REQUEST,
     CONDITIONAL_DATA,
     CONDITIONAL_DATA_FROM_MEMORY,
@@ -95,16 +104,7 @@ from .other import (
     CONDITIONAL_LINK_CONTROL_REQUEST,
     CONDITIONAL_MAX_NUMBER_OF_BLOCK_LENGTH,
     CONDITIONAL_MAX_NUMBER_OF_BLOCK_LENGTH_FILE_TRANSFER,
-    CONDITIONAL_MEMORY_ADDRESS_AND_SIZE,
-    CONDITIONAL_OPTIONAL_ADDITIONAL_PARAMETER,
-    CONDITIONAL_OPTIONAL_CHALLENGE_CLIENT,
-    CONDITIONAL_OPTIONAL_EPHEMERAL_PUBLIC_KEY_CLIENT,
-    CONDITIONAL_OPTIONAL_EPHEMERAL_PUBLIC_KEY_SERVER,
-    CONDITIONAL_OPTIONAL_NEEDED_ADDITIONAL_PARAMETER,
-    CONDITIONAL_OPTIONAL_SESSION_KEY_INFO,
     CONDITIONAL_POWER_DOWN_TIME,
-    CONDITIONAL_PROOF_OF_OWNERSHIP_CLIENT,
-    CONDITIONAL_PROOF_OF_OWNERSHIP_SERVER,
     CONDITIONAL_SECURED_DATA_TRANSMISSION_REQUEST,
     CONDITIONAL_SECURED_DATA_TRANSMISSION_RESPONSE,
     CONDITIONAL_SECURITY_ACCESS_REQUEST,
