+++ uds/translator/service_definitions/read_dtc_information.py
@@ -28,14 +28,14 @@ from ..data_record_definitions import (
     DTCS_WITH_STATUSES_AND_EXTENDED_DATA_RECORDS_DATA_LIST,
     MEMORY_SELECTION,
     MULTIPLE_DTC_AND_FAULT_DETECTION_COUNTER,
-    MULTIPLE_DTC_SEVERITY_FUNCTIONAL_UNIT_DTC_AND_STATUS,
-    MULTIPLE_DTC_SEVERITY_DTC_AND_STATUS,
     MULTIPLE_DTC_AND_SNAPSHOT_RECORD_NUMBER,
     MULTIPLE_DTC_AND_STATUS,
-    OPTIONAL_DTC_SEVERITY_FUNCTIONAL_UNIT_DTC_AND_STATUS,
+    MULTIPLE_DTC_SEVERITY_DTC_AND_STATUS,
+    MULTIPLE_DTC_SEVERITY_FUNCTIONAL_UNIT_DTC_AND_STATUS,
     OPTIONAL_DTC_AND_STATUS,
     OPTIONAL_DTC_EXTENDED_DATA_RECORD_NUMBER,
     OPTIONAL_DTC_EXTENDED_DATA_RECORDS_DATA_LIST,
+    OPTIONAL_DTC_SEVERITY_FUNCTIONAL_UNIT_DTC_AND_STATUS,
     READ_DTC_INFORMATION_SUB_FUNCTION_2013,
     READ_DTC_INFORMATION_SUB_FUNCTION_2020,
 )
